Update and rename chatsticker.py to Pcontrol.py
